requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aioredis==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
faker==20.1.0
requests==2.31.0
python-dotenv==1.0.0
typing-extensions==4.8.0